move appContext to its own file
